Changed debug print statements. initiate different callbacks based on what wifi opmode we are in. changed timestamp we are storing to a 32 bit integer

+++ ESP8266/ESP8266_SDK/sead_wifi_PROJECT/user/buffers.h
@@ -25,8 +25,7 @@ typedef struct {
 
 //exact data to populate
 typedef struct {
-   //change back to 32bit because of offset, only need 32 bit number
-	uint64_t timestamp;
+	uint32_t timestamp;
 	uint16_t wattage;
 	uint16_t vrms;
 	uint16_t irms;
